fix(core): unblock formal quality gates
