+++ apps/docs-next/content/docs/get-started/concepts/errors.mdx
@@ -63,6 +63,7 @@ ErrorCodes.AK_TOOL_FORBIDDEN        // tool execution is denied by policy
 // Memory errors
 ErrorCodes.AK_MEMORY_LOAD_FAILED          // memory.load() failed
 ErrorCodes.AK_MEMORY_SAVE_FAILED          // memory.save() failed
+ErrorCodes.AK_MEMORY_CLEAR_FAILED         // memory.clear() failed
 ErrorCodes.AK_MEMORY_DESERIALIZE_FAILED   // persisted state is corrupt
 ErrorCodes.AK_MEMORY_PEER_MISSING         // optional memory backend is not installed
 ErrorCodes.AK_MEMORY_REMOTE_HTTP          // remote memory request failed
