+++ docs/evidence/ecosystem-documentation-quality/agentskit.json
@@ -4,7 +4,7 @@
   "profileVersion": 1,
   "productId": "agentskit",
   "repo": "AgentsKit-io/agentskit",
-  "commit": "1392472aab09ef74b0728a85007fc457b9ce3293",
+  "commit": "29d0e3007c595bd12215e2778fd285ac64e7cf0e",
   "auditedOn": "2026-08-30",
   "docBridge": {
     "artifact": "docs/evidence/ecosystem-documentation-quality/doc-bridge/agentskit.json",
