+++ docs/evidence/ecosystem-documentation-quality/doc-bridge/agentskit.json
@@ -3,7 +3,7 @@
   "protocol": "agentskit.doc-bridge.attestation",
   "productId": "agentskit",
   "repo": "AgentsKit-io/agentskit",
-  "commit": "1392472aab09ef74b0728a85007fc457b9ce3293",
+  "commit": "29d0e3007c595bd12215e2778fd285ac64e7cf0e",
   "sourceMode": "commit",
   "contentDigest": "sha256:bf66923af81a3e91404e3e0971bc08f770a15923687602b94ddf427f3e460f6a",
   "command": "pnpm docs:bridge:doctor && pnpm docs:bridge:gate",
